refactor: Update import statements for AnthropicClient and OpenAIClient to use specific paths
